Changed scene transition
Changed the duck transition image. Modified various transition values to look better and work with the new asset. Fixed an issue where the wrong transition duration was being used for the second half.

Changed the scene fade transition to a duck growing and covering the screen. Modified the scene manager to add the animation. Also modified the pause menu to prevent pausing during the animation. The duck image is currently a placeholder.

Removed extra menu post processing volume