Fixed duck spawning, temp file errors and options title
Fixed the options menu title to say Options instead of Instructions

Fixed duck spawning bug and error from removed temp files

Fixed an issue where if the mouse was held during a transition screen, a large number of ducks would instantly be created after the transition was done.
Removed a reference to a removed script that was missed. The script was disabled on an active game object so didn't do anything, but still was trying to reference the missing file.