Added/Fixed screenshot system
Added a proper implementation of the screenshot system to allow pressing f2 to take a screenshot. A canvas was added that has text on it to briefly show the file name so the user knows a screenshot was taken. The reason that this feature is technically fixed was because it is built using code that was added in the last commit to the menu to make the splash screen.