Added duck collision sound effects
Added a variety of hit sound effects and a system on each high quality dynamic duck to occasionally request a hit sound effect to play. These requests are managed by a sound effect manager so that only a limited number of sound effects can play at any given moment. When the user places only a single duck, it is marked as being important by changes to the spawning and dynamic queue scripts. This makes the sound effect code on the duck bypass the sound effect manager to always play the sound and allows the volume to be slightly increased for these important sounds.