Improved sound effects
Changed how sound effects are handled. When a duck has a high enough velocity impact and a couple other smaller conditions are met, it requests to the sound manager to play a hit sound. It will send a reference to itself and the volume of the sound it will play. The sound effect manager will pick the duck that will make the most noise and have it play a sound. It may have the duck play multiple sounds if the framerate is low to maintain the correct amount of noise. If the duck that was spawned was the from a click, it is marked as important and bypasses the sound manager, simply playing the sound.
Also added a debug feature where pressing keys will change the framerate to easily test how things respond to lag.