Added an audio setting screen and fixed misc issues
Added an audio settings screen to the option screen on the main menu. Added two sliders for volume along with fixing issues to do with camera rotation and screen transitions. Added functionality to change the volume level of all elements based on the settings chosen.
Fixed a sound issue where newly spawned ducks would create a hit sound due to spawning inside another duck.
Fixed an issue where the shadow blocker for fair settings was not copied to the main menu.